pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sandbox-mcp"
version = "0.1.0"
description = "General-purpose Python execution sandbox with FastMCP tools, artifact capture, Manim rendering, and web demo workflows"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["mcp", "sandbox", "python", "manim", "artifacts", "fastmcp"]
dependencies = [
    "aiohttp>=3.9.0",
    "fastmcp>=2.10.5",
    "ffmpeg-python==0.2.0",
    "ipython>=9.4.0",
    "manim==0.19.0",
    "numpy>=1.26.4",
    "matplotlib>=3.8.4",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "scikit-learn==1.5.2",
    "seaborn==0.12.0",
]

[project.urls]
Homepage = "https://github.com/scooter-lacroix/sandbox-mcp"
Documentation = "https://github.com/scooter-lacroix/sandbox-mcp/blob/main/README.md"
Repository = "https://github.com/scooter-lacroix/sandbox-mcp"
Issues = "https://github.com/scooter-lacroix/sandbox-mcp/issues"

[project.scripts]
sandbox-mcp = "sandbox.mcp_sandbox_server_stdio:main"
sandbox-mcp-http = "sandbox.mcp_sandbox_server:main"
sandbox-server = "sandbox.mcp_sandbox_server:main"
sandbox-server-stdio = "sandbox.mcp_sandbox_server_stdio:main"

[project.optional-dependencies]
web = [
    "flask>=3.0.0",
    "streamlit>=1.30.0",
]
dev = [
    "black>=24.0.0",
    "mypy>=1.8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sandbox"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
